move flaky test to repeatable
